requirements.txt
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.30
pandas-datareader>=0.10.0
pyyaml>=6.0
pyarrow>=13.0.0
fastparquet>=2023.8.0
feedparser>=6.0.10
newspaper3k>=0.2.8
beautifulsoup4>=4.12.2
lxml>=4.9.3

psycopg2-binary>=2.9.7
SQLAlchemy>=2.0.0

pymongo>=4.5.0
python-dateutil>=2.8.2